Added variables for ball attributes
The size, color, and speed of balls are now all linked to global variables. The values of these variables will be changed later with a GUI, giving the user full control over the initial state of the balls.

Eventually the ball will appear at the point where the mouse clicks, but a framework for this has not been implemented yet.