Fixed Firefox Version
for some reason <orient="vertical"> was breaking the gravity slider in firefox, but it wasn't adding anything in other browsers since I am using js to rotate the slider. 

To be honest I am not even sure that <orient="vertical"> is a real thing.. 

Will test ie when I am able.